moves explore to src/routes
