+++ src/routes/_components/GLAMFooterContent.svelte
