+++ src/routes/_components/KeySelectionControl.svelte
