+++ src/routes/_components/ProbeDetails.svelte
