+++ src/routes/_components/ProbeKeySelector.svelte
@@ -1,9 +1,9 @@
 <script>
-  import FloatingMenu from '../../../components/menu/FloatingMenu.svelte';
-  import MenuList from '../../../components/menu/MenuList.svelte';
-  import MenuListItem from '../../../components/menu/MenuListItem.svelte';
-  import DownCarat from '../../../components/icons/DownCarat.svelte';
-  import { tooltip as tooltipAction } from '../../../components/utils/tooltip';
+  import FloatingMenu from '../../components/menu/FloatingMenu.svelte';
+  import MenuList from '../../components/menu/MenuList.svelte';
+  import MenuListItem from '../../components/menu/MenuListItem.svelte';
+  import DownCarat from '../../components/icons/DownCarat.svelte';
+  import { tooltip as tooltipAction } from '../../components/utils/tooltip';
 
   export let options;
   export let currentKey;
