+++ src/routes/_components/ProportionMetricTypeControl.svelte
