+++ src/routes/_components/ReferenceSymbol.svelte
