+++ src/routes/_components/defaultview/MarketingBlock.svelte
