+++ src/routes/_components/defaultview/RandomProbePlaceholder.svelte
