+++ src/routes/_components/defaultview/sm-logic.js
