+++ src/routes/_components/errors/DataError.svelte
