+++ src/routes/explore/_view/ComparisonSummary.svelte
