+++ src/routes/explore/_view/ProbeExplorer.svelte
@@ -8,11 +8,11 @@ import DistributionComparison from './DistributionComparison.svelte';
 import ComparisonSummary from './ComparisonSummary.svelte';
 import Violin from '../../../components/data-graphics/elements/Violin.svelte';
 
-import { explorerComparisonSmallMultiple } from '../utils/constants';
+import { explorerComparisonSmallMultiple } from '../../../utils/constants';
 
-import { formatBuildIDToDateString } from '../utils/formatters';
+import { formatBuildIDToDateString } from '../../../utils/formatters';
 
-import { histogramSpring } from '../utils/animation';
+import { histogramSpring } from '../../../utils/animation';
 
 import { extractBinValues } from '../../../utils/probe-utils';
